requirements.txt
streamlit
langchain
pymupdf
PyPDF2
pypdf
python-dotenv
tiktoken
chromadb
openai